+++ example-views/yiisoft/yii2-app/layouts/left.php
@@ -9,7 +9,7 @@
     <div class="sidebar">
 
         <nav class="mt-2">
-            <?= dmstr\adminlte\widgets\Menu::widget(
+            <?= memouk\adminlte\widgets\Menu::widget(
                 [
                     'options' => ['class' => 'nav nav-pills nav-sidebar flex-column', 'data-widget' => 'treeview'],
                     'items' => [