+++ CHANGELOG.md
@@ -3,7 +3,7 @@ CHANGELOG
 
 ## Unreleased (3.0)
  * updated example views, widgets and assets to AdminLTE 3
- * changed namespace to `dmstr/adminlte`
+ * changed namespace to `memouk/adminlte`
  * added `iconClassType` to `Menu` widget to support different types of FontAwesome 5 icons
 
 ## 2.6.0